+++ .github/workflows/ci_testing.yml
@@ -51,7 +51,7 @@ jobs:
 
     - name: Tests
       run: |
-        python -m pytest . -v --cov=kaggle_plantpatho --junitxml=junit/test-results-${{ runner.os }}-${{ matrix.python-version }}.xml
+        python -m pytest . -v --cov=kaggle_plantpathology --junitxml=junit/test-results-${{ runner.os }}-${{ matrix.python-version }}.xml
 
     - name: Upload pytest test results
       uses: actions/upload-artifact@v2
