+++ MANIFEST.in
@@ -5,7 +5,7 @@ recursive-exclude __pycache__  *.py[cod] *.orig
 
 # Include the README and CHANGELOG
 include *.md
-recursive-include kaggle_plantpatho *.md
+recursive-include kaggle_plantpathology *.md
 recursive-include assets *.png *.jpg
 
 # Include the license file
