+++ tests/plant_pathology/test_data.py
@@ -3,7 +3,7 @@ import os
 import numpy
 import pytest
 
-from kaggle_plantpatho.data import PlantPathologyDataset, PlantPathologyDM, PlantPathologySimpleDataset
+from kaggle_plantpathology.data import PlantPathologyDataset, PlantPathologyDM, PlantPathologySimpleDataset
 from torch import Tensor
 
 _PATH_HERE = os.path.dirname(__file__)
@@ -46,8 +46,8 @@ _TEST_LABELS_BINARY = [
 @pytest.mark.parametrize("phase", ["train", "valid"])
 def test_dataset(data_cls, labels, phase, root_path=_PATH_HERE):
     dataset = data_cls(
-        df_data=os.path.join(root_path, "data", "train.csv"),
-        path_img_dir=os.path.join(root_path, "data", "train_images"),
+        df_data=os.path.join(root_path, "../data_plant-pathology", "train.csv"),
+        path_img_dir=os.path.join(root_path, "../data_plant-pathology", "train_images"),
         split=1.0 if phase == "train" else 0.0,
         mode=phase,
     )
@@ -69,7 +69,7 @@ def test_dataset(data_cls, labels, phase, root_path=_PATH_HERE):
 def test_datamodule(simple, balance, root_path=_PATH_HERE):
     dm = PlantPathologyDM(
         path_csv="train.csv",
-        base_path=os.path.join(root_path, "data"),
+        base_path=os.path.join(root_path, "../data_plant-pathology"),
         simple=simple,
         split=0.6,
         balancing=balance,
